Change gear to use a multiply instead of a table lookup.
A table lookup uses 2K of L1 cache which can be slower than a multiply in the
broader context of an application using that L1 cache. Using a multiply
requires no L1 cache, and can hash as good or better than a table lookup. This
is an idea taken from testing and analysis done at;

* https://github.com/dbaarda/rollsum-tests/blob/master/RESULTS.rst